fix: remove random generate heights from VRTs for spacings
Also, move all VRTs into one suite for spacings to save on some snaps. #snaps
